pyproject.toml
[project]
name = "script-rosseta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
]
